sequence so incremental db updates can depend on monolithic db

+++ src/sr/db/tree_update.cpp
@@ -13,9 +13,15 @@ DeleteCommand::DeleteCommand( const TreeZone &target_ ) :
 
 void DeleteCommand::Execute( const ExecKit &kit ) const
 {
-	kit.x_tree_db->Delete( target );    
+	if( kit.monolithic_is_valid ) // Incrementals might depend on monolithics
+    {
+        kit.x_tree_db->Delete( target );    
+    }
     
-    target.GetBase().ClearXPtr();
+    if( !kit.monolithic_is_valid ) // Monolithics might depend on tree update
+    {
+        target.GetBase().ClearXPtr();
+    }
 }	
 
 
@@ -29,11 +35,17 @@ InsertCommand::InsertCommand( const TreeZone &target_, const FreeZone &new_zone_
 void InsertCommand::Execute( const ExecKit &kit ) const
 {
     XLink base_xlink = target.GetBase();
-    ASSERT( !base_xlink.GetChildX() );
-	base_xlink.SetXPtr( new_zone.GetBase() );
+    if( !kit.monolithic_is_valid ) // Monolithics might depend on tree update
+    {
+        ASSERT( !base_xlink.GetChildX() );
+        base_xlink.SetXPtr( new_zone.GetBase() );
+    }
     
-    TreeZone nt( base_xlink, new_zone );
-    kit.x_tree_db->Insert( nt );      
+    if( kit.monolithic_is_valid ) // Incrementals might depend on monolithics
+    {
+        TreeZone nt( base_xlink, new_zone );
+        kit.x_tree_db->Insert( nt );      
+    }
 }
